Changed helloworld help article to use plain RWT
The launch step in the helloworld article did not work because the RWT
launcher starts in JEE mode, which does not support the main loop.

The helloworld article should not lead to a first workbench application
anyway. Reworked to use only RWT and link to next part of application
development.