Initialize target platform in tests to prevent errors
When running tests with tycho, the default PDE target platform is
empty because PDE initializes the target by searching the path
${eclipse_home} for plug-ins.

The tests for RAPLaunchDelegate indirectly call code in PDE that
relies on the system bundle to be present in the target platform.

Introduce a utility that initializes the target platform from the
bundles in the running OSGi platform. The code in this utility
originates from the GEF project (see bug 422952).

Fix bug 422952: [Tools][releng] NPE when running unit tests with tycho
https://bugs.eclipse.org/bugs/show_bug.cgi?id=422952