Fix doc build for Windows and for spaces in path
In the pom.xml, I've split a very long parameter list into separate lines
by escaping newlines. It turned out that this practice does not work on
Windows: all subsequent are ignored. So we have to live with this one
very long line in the pom.

Moreover, surround paths with quotes to let the build work when the git
repository path contains spaces.

Change-Id: Ib1f10f600c230827015bd2444e54696340389156