Add build qualifier to repository zip
Use the tycho property ${buildQualifier} in the repository zip
filename. In our build, this qualifier is generated by the tycho-
buildtimestamp-jgit plug-in. For eclipse-repository projects, it
is apparently bound to the included feature's build qualifiers.

With this change, the build script becomes mostly redundant and
should be deleted once the hudson build doesn't need it anymore.

Change-Id: I25d12de9b7d5a9225324d197815946729e4166d2