Add common top-level files to Git repository
Move the module specification from the parent pom.xml in
releng/org.eclipse.rap.tools.build to a new pom.xml in the root
directory of the Git repository.

Fix error with target definition that was a consequence of the move of
the main pom.xml by extracting it into its own eclipse-target-definition
module.

Add a CONTRIBUTING.md that describes the basics of contributing to an
Eclipse project, and add a README.md with a more detailed description on
how the projects are layed out in the Git repository, how to build the
project, the license, and how to interact with the project.

Bug 424687: Add common top-level files to git repositories
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424687


Change-Id: I603c778aa56a0dcd3c7b78c384d0f233cf63ea15
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>