Move license feature repo definition from pom.xml to target definition
RAP Tools uses a target definition file for the build instead of a list
of repositories in the pom.xml, therefore the best place for the new
repository definition is the existing target definition file. Moreover,
this is a hard dependency on a specific version (1.0.1.v20140414-1359)
of the license feature.

Change-Id: I0ede21e1b000eb2578fc5d948eec08c743557a80
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>