Upgrade RAP Tools build to Eclipse Tycho 1.0.0
Upgrade Eclipse Tycho for building to version 1.0 following the old
tradition of using the latest published build for building which brings
it to the same level as the Eclipse Platform build for Oxygen.

The release notes do *not* list any incompatibilities that would affect
the RAP Tools build:

https://wiki.eclipse.org/Tycho/Release_Notes/0.26
https://wiki.eclipse.org/Tycho/Release_Notes/1.0.0

One advantage is the usage of checksums even for external bundles.

Change-Id: I722be1506ecf3b5394ea7775e85b3f4da49b4c6f
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>