Fix build errors when generating API documentation
The API documentation is generated as javadoc as part of a multi-layer
wrapped build process: javadoc is running inside an Ant script started
from an Eclipse which is controlled by the Tycho Eclipse Runner which
is part of the Maven build process.

Unfortunately, the Tycho Eclipse Runner allows by default a JavaSE-1.7
as execution environment, and uses it if available somewhere on the
build system like Jenkins.

Explicitly define the execution environment, and update the Eclipse
version accordingly.

Change-Id: I94fbf7fa5f2aedc5859a1e6af79ad28e3e6bffe0
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>