Work around p2 / OSGi capability requirements issues in Jetty
For the time being the provided Jetty 10 bundles define some hard
requirements to OSGi capabilities, although they are not in use by RAP
Tools.

Work around this by providing virtual capabilities in the RAP Tools
launch bundle until these requirements are marked as optional by Jetty.
In this case we add the missing osgi.serviceloader capability to
org.eclipse.jetty.webapp.Configuration which is not required in the RAP
Tools environment.

Change-Id: I26e7c5db00807fcc73eae4e4aff2fb0b6f56f4b9
Signed-off-by: Markus Knauer <mknauer@eclipsesource.com>