Let Tycho p2 manage external dependencies
In earlier times the build dependencies to external bundles of Eclipse
Platform, Jetty, and license bundles was managed by a target definition
file as it is in use by Eclipse PDE. One problem of this approach is
that the RAP target definition relied on exact version specifications
which are painful to keep in sync, the new upcoming problem is the
strategy of Eclipse Platform to move away from coherent features to
bundles which leads to a mixture of bundle & feature installable units
in the RAP target definition.

Replace the target definition file in the build with p2 repository
location URLs as it is done in RAP Runtime since the beginning.