Load all dependencies in default RAP Tools target definitions
When loading a default RAP target definition, e.g. as part of the
creation of a wizard-generated plug-in, a predefined target definition
is loaded.

This target definition used to be resolved by the 'slicer' 'includeMode'
of p2 which uses strict dependencies only.

Use the 'includeMode' 'planner' instead which includes dependencies.