Upgrade to Eclipse Tycho 3.0.4
The old Tycho version 2.7.5 had to be replaced with the latest Tycho
version 3.0.4 to fulfill Java 17 requirements of third party
dependencies in the Eclipse Platform.

Following the migration guide at [1] this includes removal of the pack
200 post-processing, migration of the Git timestamp provider, and
setting the archive format for Linux back to ZIP.

Also update the Eclipse version and its runtime arguments that are used
for dynamically creating the HTML documentation at runtime of the build.

[1] https://github.com/eclipse-tycho/tycho/blob/tycho-3.0.4/RELEASE_NOTES.md#migration-guide-2x---3x