Switch to new Jakarta annotation API bundle in e4 template
The e4 RAP project template uses the `javax.annotation` packages which
used to be provided by a bundle with the same bundle symbolic name. With
the new release the very same namespace has been renamed and is now
provided by a bundle `jakarta.annotation-api`.

Update the launch configuration, and the wizard that creates the
MANIFEST.MF file.

Also add and update missing bundles to all created launch
configurations.